change datatype in accuracy_percentage column for stats tables

+++ src/core/models/statistics_data.py
@@ -47,7 +47,7 @@ class BaseStats:
     )
 
     symbols_per_minute: Mapped[int] = mapped_column()
-    accuracy_percentage: Mapped[int] = mapped_column()
+    accuracy_percentage: Mapped[float] = mapped_column()
 
 
 class StandardModeStats(
